refactor: small fixes, reuse login form, delete profile routes func
