+++ client/src/app/router/index.ts
@@ -1,3 +1,2 @@
 export { Router } from './ui/router';
-export { getProfileRoutes } from './lib/get-profile-routes';
 export * from './model/routes.types';
