--- client/src/app/router/lib/get-profile-routes.tsx
@@ -1,4 +1,0 @@
-import { EnumRoutesName } from 'shared/config';
-
-export const getProfileRoutes = () =>
-  Object.values(EnumRoutesName).filter((route) => route === EnumRoutesName.ORDERS);
