+++ client/src/features/auth/index.ts
@@ -1,3 +1,3 @@
 export * from './model/auth.store';
 export * from './model/auth.types';
-export * from './model/login.constants';
+export * from './model/auth.constants';
