+++ client/src/features/auth/model/auth.constants.ts
