--- client/src/shared/ui/order-card/index.ts
@@ -1,2 +1,0 @@
-export * from './model/order-card.types';
-export { OrderCard } from './ui/order-card';
