+++ client/src/widgets/layout/ui/layout.tsx
@@ -4,7 +4,7 @@ import { Header } from './header';
 import { Footer } from './footer';
 import { StyledContainer, StyledContent } from './layout.styles';
 
-export const Layoutponent = (props: ILayoutProps) => {
+export const LayoutСomponent = (props: ILayoutProps) => {
   const { children } = props;
 
   return (
