+++ client/src/widgets/login/model/login-form.types.ts
@@ -1,0 +1,3 @@
+export interface ILoginForm {
+  formType: 'email' | 'phone';
+}
