+++ client/src/widgets/register/index.ts
@@ -1,1 +1,1 @@
-export { RegisterForm } from './ui/register-form';
+export { RegisterForm } from './ui/register';
