+++ client/src/widgets/register/ui/register.styles.ts
